register page, log in page and user account page

+++ routers/posts.py
@@ -8,7 +8,7 @@ from sqlalchemy.orm import selectinload
 import models
 from database import get_db
 from schemas import PostCreate, PostResponse, PostUpdate
-
+from auth import CurrentUser
 router = APIRouter()
 
 
@@ -31,21 +31,12 @@ async def get_posts(db: Annotated[AsyncSession, Depends(get_db)]):
     response_model=PostResponse,
     status_code=status.HTTP_201_CREATED,
 )
-async def create_post(post: PostCreate, db: Annotated[AsyncSession, Depends(get_db)]):
-    result = await db.execute(
-        select(models.User).where(models.User.id == post.user_id)
-    )
-    user = result.scalars().first()
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="User not found",
-        )
+async def create_post(post: PostCreate, current_user: CurrentUser, db: Annotated[AsyncSession, Depends(get_db)]):
 
     new_post = models.Post(
         title=post.title,
         content=post.content,
-        user_id=post.user_id,
+        user_id=current_user.id,
     )
     db.add(new_post)
     await db.commit()
@@ -69,7 +60,7 @@ async def get_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]):
 ## full update post using post id
 # note: ASYNC
 @router.put("/{post_id}", response_model=PostResponse)
-async def update_post_full(post_id: int, post_data: PostCreate, db: Annotated[AsyncSession, Depends(get_db)]):
+async def update_post_full(post_id: int, current_user: CurrentUser, post_data: PostCreate, db: Annotated[AsyncSession, Depends(get_db)]):
     result = await db.execute(
         select(models.Post)
         .where(models.Post.id == post_id)
@@ -78,18 +69,11 @@ async def update_post_full(post_id: int, post_data: PostCreate, db: Annotated[As
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found")
 #     Check if user id of the retrieved post == post_data's user_id, in other word, check if the user exists
-
-    if post.user_id != post_data.user_id:
-        result = db.execute(select(models.User).where(models.User.id == post_data.user_id))
-        user = result.scalars().first()
-        if not user:
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail="User not found",
-            )
+    if post.user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to update this post")
 
     post.title = post_data.title
-    post.user_id = post_data.user_id
+    post.user_id = post_data.current_user.id
     post.content = post_data.content
 
     await db.commit()
@@ -99,12 +83,15 @@ async def update_post_full(post_id: int, post_data: PostCreate, db: Annotated[As
 ## partially update post using post id
 # note: ASYNC
 @router.patch("/{post_id}", response_model=PostResponse)
-async def update_post_partial(post_id: int, post_data: PostUpdate, db: Annotated[AsyncSession, Depends(get_db)]):
+async def update_post_partial(post_id: int, current_user: CurrentUser, post_data: PostUpdate, db: Annotated[AsyncSession, Depends(get_db)]):
     result = await db.execute(select(models.Post).where(models.Post.id == post_id))
     post = result.scalars().first()
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found")
 #     Check if user id of the retrieved post == post_data's user_id, in other word, check if the user exists
+
+    if post.user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to update this post")
 
     update_post = post_data.model_dump(exclude_unset=True)
     for (field, value) in update_post.items():
@@ -116,10 +103,13 @@ async def update_post_partial(post_id: int, post_data: PostUpdate, db: Annotated
 ## delete post
 # note:ASYNC
 @router.delete("/{post_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]):
+async def delete_post(post_id: int, current_user: CurrentUser, db: Annotated[AsyncSession, Depends(get_db)]):
     result = await db.execute(select(models.Post).where(models.Post.id == post_id))
     post = result.scalars().first()
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found")
+    if post.user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to delete this post")
+
     await db.delete(post)
     await db.commit()
